Rename UserSettingsWatcher to CachedSettingWatcher
